Completely remove reversibility
The reversibility system I had in place was perfect apart from a single flaw that I could not get around. When acting on a process like (a+(b|c)), acting on b or c will have to set the |'s past life to the summation, but the actionprefix pastlife will remain as just an actionprefix. This can be seen by running a program like that in the previous commit.

I need to do some major structural changes in order to make this work

Signed-off-by: GMX <pbrowning7@gmail.com>